add responses to topics

+++ public/js/topic.js
@@ -3,17 +3,18 @@ const topic_create_form_handler = async (event) => {
   event.preventDefault();
 
   const subject = document.querySelector('#topic_subject').value.trim();
-  const contents = document.querySelector('#topic_contents').value.trim();
+  const content = document.querySelector('#topic_content').value.trim();
 
-  if (subject && contents) {
+  if (subject && content) {
     const response = await fetch('/api/topic', {
       method: 'POST',
-      body: JSON.stringify({ subject, contents }),
+      body: JSON.stringify({ subject, content }),
       headers: { 'Content-Type': 'application/json' },
     });
 
     if (response.ok) {
-      document.location.replace('/');
+      /* Onn success, refresh the dashboard to show the new topic.  */
+      document.location.replace('/dashboard');
     } else {
       alert('Failed to create topic.');
     }
@@ -21,7 +22,7 @@ const topic_create_form_handler = async (event) => {
 };
 
 document
-  .querySelector('.topic_create_form')
+  .querySelector('#topic_create_form')
   .addEventListener('submit', topic_create_form_handler);
 
 /* Delete a topic */
@@ -29,12 +30,21 @@ document
 async function topic_delete_handler(event) {
   const the_button = event.target;
   const topic_id = the_button.dataset.id;
+  /* If there is no data-id onn the target, we must be
+   * seeing a click from somewhere else in the
+   * page.  Ignore it.  */
+  if (topic_id == null) {
+    return;
+  }
+  
   const response = await fetch('api/topic/' + topic_id, {
     method: 'DELETE',
     headers: { 'Content-Type': 'application/json' },
   });
   if (response.ok) {
-    document.location.replace('/profile');
+    /* On success, refresh the dashboard to show that the topic
+     * is no longer present.  */
+    document.location.replace('/dashboard');
   } else {
     alert('Failed to delete topic.');
   }
